+++ Backend/Threads/Handles/ConstructBehaviorFeatureCheckLoop.cs
@@ -5,9 +5,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Mod.DynamicEncounters.Features.Interfaces;
 using Mod.DynamicEncounters.Helpers;
-using Mod.DynamicEncounters.Threads;
 
-namespace Mod.DynamicEncounters;
+namespace Mod.DynamicEncounters.Threads.Handles;
 
 public class ConstructBehaviorFeatureCheckLoop(IThreadManager tm, CancellationToken ct) :
     ThreadHandle(ThreadId.BehaviorFeatureCheck, tm, ct)
